// web/src/features/downloads/types.ts
/*
Copyright (C) 2023-2026 QuantumNous

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License
along with this program. If not, see <https://www.gnu.org/licenses/>.

For commercial licensing, please contact support@quantumnous.com
*/
export type DownloadPlatform = 'windows' | 'macos' | 'linux' | 'android'

export type ClientDownloadAsset = {
  name: string
  platform: DownloadPlatform
  arch: string
  format: string
  size: number
  url: string
  mirror_url?: string
  sha256?: string
}

export type ClientDownloadRelease = {
  version: string
  published_at: string
  notes: string
  release_url: string
  assets: ClientDownloadAsset[]
}

export type ClientDownloadsResponse = {
  success: boolean
  message?: string
  data?: ClientDownloadRelease
}